https://drive.google.com/file/d/1CQsABiCxoRWmEAbg13XiXjGlt34oKz1d/view?usp=sharing
I recently participated in Tata's job simulation on the Forage platform, and it was incredibly useful to understand what it might be like to participate on a consulting team at Tata.
I worked on a project to help a retail client review their revenue data to prepare for the future. I created data visuals, practiced using PowerBI to represent information, and built my presentation skills in a real-world context.